Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
google-auth>=2.25.0
google-cloud-aiplatform>=1.38.0
python-multipart>=0.0.6
pymongo>=4.6.0
boto3>=1.34.0